refactor: complete shared type inventory
