+++ src/App.tsx
@@ -65,7 +65,7 @@ async function ensureRivenWindow(wx: number, wy: number, wh: number): Promise<{ 
 import { getCurrentWindow, availableMonitors, LogicalSize } from "@tauri-apps/api/window";
 
 import { ImgCacheDirContext } from "./ImgCacheDir";
-import Foundry, { FoundryFilters, FOUNDRY_FILTERS_DEFAULT } from "./Foundry";
+import Foundry, { FOUNDRY_FILTERS_DEFAULT } from "./Foundry";
 import CacheStatusChip from "./CacheStatusChip";
 import MarketHelper, { MARKET_FILTERS_DEFAULT } from "./MarketHelper";
 import RelicHelper, { RELIC_FILTERS_DEFAULT } from "./RelicHelper";
@@ -75,7 +75,7 @@ import RelicPickOverlay from "./RelicPickOverlay";
 import TimerHelper, { fmtMs } from "./TimerHelper";
 import { useWorldState } from "./worldstate";
 import { notify } from "./notify";
-import { collectNewMatches, type SeenFissures } from "./fissureAlerts";
+import { collectNewMatches } from "./fissureAlerts";
 import Statistics from "./Statistics";
 import Syndicates from "./Syndicates";
 import Weapons from "./Weapons";
@@ -87,15 +87,16 @@ import ChangeLog from "./ChangeLog";
 import InventoryGrid from "./InventoryGrid";
 import InventoryBatchPreview from "./InventoryBatchPreview";
 import InventoryToolbar from "./InventoryToolbar";
-import { INVENTORY_FILTERS_DEFAULT, type InventoryFilters } from "./InventoryFilters";
+import { INVENTORY_FILTERS_DEFAULT } from "./InventoryFilters";
+import type { FoundryFilters, InventoryFilters } from "./types/filters";
 import type { ViewMode } from "./types/ui";
 import { formatUnixTime } from "./formatters";
 import type { ArchonShard, CatalogItem, CraftingJob, InventoryItem } from "./types/items";
 import type { ChangeLogEntry, InventoryUpdate, ModCopy } from "./types/inventory";
 import type { RivenAnalysis, RivenAnalysisUpdate, RivenStat } from "./types/rivens";
 import type { ClockFormat, FissureWatch, FoundryPageSize, RelicOverlayPriority, RelicPickLines, RelicPickPriority, RelicRefinement, SettingsSnapshot } from "./types/settings";
+import type { SeenFissures } from "./types/worldstate";
 import type { TradeCompletedEvent } from "./types/trades";
-export type { InventoryItem } from "./types/items";
 import "./App.css";
 import "./images.css";
 import "./InventoryGrid.css";
