+++ src/InventoryFilters.ts
@@ -1,16 +1,4 @@
-export type InventorySortMode = "qty-desc" | "qty-asc" | "name-asc" | "name-desc" | "recent";
-
-export interface InventoryFilters {
-  category: string;
-  search: string;
-  filterOwned: boolean;
-  filterRecent: boolean;
-  filterPrime: boolean;
-  filterVaulted: boolean;
-  filterUnvaulted: boolean;
-  filterRank: number | "unranked" | null;
-  sortMode: InventorySortMode;
-}
+import type { InventoryFilters } from "./types/filters";
 
 export const INVENTORY_FILTERS_DEFAULT: InventoryFilters = {
   category: "all",
