+++ src/types/worldstate.ts
@@ -1,3 +1,5 @@
+import type { FissureVariant } from "./settings";
+
 export interface WsCycle { expiry: string; }
 export interface WsCetus extends WsCycle { isDay: boolean; }
 export interface WsVallis extends WsCycle { isWarm: boolean; }
@@ -14,6 +16,8 @@ export interface WsPrimeResurgence { expiry: string; activation: string; active:
 export interface WsNight { expiry: string; season: number; active: boolean; }
 export interface WsFissure { id: string; expiry: string; node: string; missionType: string; enemy: string; tier: string; tierNum: number; isStorm: boolean; isHard: boolean; active: boolean; }
 export interface WsStorm { id: string; expiry: string; node: string; missionType: string; enemy: string; tier: string; tierNum: number; active: boolean; }
+export type MatchedFissure = { f: WsFissure | WsStorm; variant: FissureVariant };
+export type SeenFissures = Map<string, Set<string>>;
 export interface WsAlert { id: string; expiry: string; missionType: string; faction: string; node: string; rewardItem?: string; rewardCredits: number; }
 export interface WsInvasion { id: string; node: string; attacker: string; defender: string; attReward: string; defReward: string; pct: number; }
 export interface WsDarvo { expiry: string; item: string; discount: number; originalPrice: number; salePrice: number; amountTotal: number; amountSold: number; }
